+++ src/main/java/com/xtuer/mapper/EnrollmentMapper.java
@@ -5,8 +5,8 @@ import org.apache.ibatis.annotations.Param;
 
 public interface EnrollmentMapper {
     //根据认定id查询认定历史表
-    EnrollmentForm findHistoryValidByRegisterId(@Param("registerId") int registerId);
+    EnrollmentForm findHistoryValidByRegisterId(@Param("registerId") long registerId);
     int insertEnrollment(EnrollmentForm form);
     //根据认定Id查询认定正式表
-    EnrollmentForm findRegistrationByRegisterId(@Param("registerId") int registerId);
+    EnrollmentForm findRegistrationByRegisterId(@Param("registerId") long registerId);
 }
